+++ src/components/sidebar/Sidebar.jsx
@@ -53,8 +53,8 @@ import style from './Sidebar.module.css'
 export default function Sidebar({ collapsed = true, collapsable = false }) {
   const user = useSelector((state) => state.user)
   const lobby = useSelector((state) => state.lobby)
-  const match = useSelector((state) => state.matchmaking.match)
-  const preMatch = useSelector((state) => state.matchmaking.preMatch)
+  const match = useSelector((state) => state.match)
+  const preMatch = useSelector((state) => state.preMatch)
   const notifications = useSelector((state) => state.notifications)
   const invites = useSelector((state) => state.invites)
   const friendListOpenByApp = useSelector((state) => state.app.friendListOpen)