+++ src/components/sidebar/Sidebar.stories.jsx
@@ -7,8 +7,9 @@ import AppReducer from '@slices/AppSlice'
 import FriendReducer from '@slices/FriendSlice'
 import InviteReducer from '@slices/InviteSlice'
 import LobbyReducer from '@slices/LobbySlice'
-import MatchmakingReducer from '@slices/MatchmakingSlice'
+import MatchReducer from '@slices/MatchSlice'
 import NotificationReducer from '@slices/NotificationSlice'
+import PreMatchReducer from '@slices/PreMatchSlice'
 import UserReducer from '@slices/UserSlice'
 
 export default {
@@ -218,10 +219,10 @@ export const Default = {
       })
     )
 
-    const matchmaking = {
-      preMatch: null,
-      match: props.isInMatch,
-    }
+    const match = props.isInMatch
+
+    const preMatch = null
+
     const app = {
       toasts: [],
       friendListOpen: false,
@@ -241,10 +242,11 @@ export const Default = {
         user: UserReducer,
         notifications: NotificationReducer,
         invites: InviteReducer,
-        matchmaking: MatchmakingReducer,
+        match: MatchReducer,
         app: AppReducer,
         friends: FriendReducer,
         lobby: LobbyReducer,
+        preMatch: PreMatchReducer,
       },
       preloadedState: {
         user,
@@ -252,8 +254,9 @@ export const Default = {
         notifications,
         invites,
         app,
-        matchmaking,
+        match,
         lobby,
+        preMatch,
       },
     })